+++ VisualStudio/Hengband/Hengband.vcxproj
@@ -559,7 +559,7 @@
     <ClCompile Include="..\..\src\save\info-writer.cpp" />
     <ClCompile Include="..\..\src\save\item-writer.cpp" />
     <ClCompile Include="..\..\src\save\monster-entity-writer.cpp" />
-    <ClCompile Include="..\..\src\save\monster-writer.cpp" />
+    <ClCompile Include="..\..\src\save\lore-writer.cpp" />
     <ClCompile Include="..\..\src\save\player-writer.cpp" />
     <ClCompile Include="..\..\src\save\save-util.cpp" />
     <ClCompile Include="..\..\src\object-activation\activation-others.cpp" />
@@ -1370,7 +1370,7 @@
     <ClInclude Include="..\..\src\save\info-writer.h" />
     <ClInclude Include="..\..\src\save\item-writer.h" />
     <ClInclude Include="..\..\src\save\monster-entity-writer.h" />
-    <ClInclude Include="..\..\src\save\monster-writer.h" />
+    <ClInclude Include="..\..\src\save\lore-writer.h" />
     <ClInclude Include="..\..\src\save\player-writer.h" />
     <ClInclude Include="..\..\src\save\save-util.h" />
     <ClInclude Include="..\..\src\object-activation\activation-others.h" />
