+++ VisualStudio/Hengband/Hengband.vcxproj.filters
@@ -1613,7 +1613,7 @@
     <ClCompile Include="..\..\src\save\item-writer.cpp">
       <Filter>save</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\save\monster-writer.cpp">
+    <ClCompile Include="..\..\src\save\lore-writer.cpp">
       <Filter>save</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\save\info-writer.cpp">
@@ -4376,7 +4376,7 @@
     <ClInclude Include="..\..\src\save\item-writer.h">
       <Filter>save</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\save\monster-writer.h">
+    <ClInclude Include="..\..\src\save\lore-writer.h">
       <Filter>save</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\save\info-writer.h">
