+++ src/save/lore-writer.h
